DBZ-988 Break down Unwrap TestIT Abstract
This allows other tests to also take advantages of the testing structure
